+++ test/demuxer_test.exs
@@ -26,11 +26,13 @@ defmodule Membrane.WebM.DemuxerTest do
       ]
 
       {{:ok, spec: %ParentSpec{children: children, links: links}},
-       %{output_dir: options.output_dir}}
+       %{output_dir: options.output_dir, track_id_to_file: options.track_id_to_output_file}}
     end
 
     @impl true
     def handle_notification({:new_track, {track_id, track_info}}, :demuxer, _context, state) do
+      output_file = state.track_id_to_file[track_id]
+
       cond do
         track_info.codec == :opus ->
           children = %{
@@ -39,7 +41,7 @@ defmodule Membrane.WebM.DemuxerTest do
               serial_number: 4_210_672_757
             },
             {:sink, track_id} => %Membrane.File.Sink{
-              location: Path.join(state.output_dir, "#{track_id}.ogg")
+              location: Path.join(state.output_dir, output_file)
             }
           }
 
@@ -58,7 +60,7 @@ defmodule Membrane.WebM.DemuxerTest do
           children = %{
             {:serializer, track_id} => Membrane.Element.IVF.Serializer,
             {:sink, track_id} => %Membrane.File.Sink{
-              location: Path.join(state.output_dir, "#{track_id}_#{codec}.ivf")
+              location: Path.join(state.output_dir, output_file)
             }
           }
 
@@ -70,23 +72,50 @@ defmodule Membrane.WebM.DemuxerTest do
           ]
 
           {{:ok, spec: %ParentSpec{children: children, links: links}}, state}
+
+        track_info.codec == :h264 ->
+          codec = Atom.to_string(track_info.codec)
+
+          children = %{
+            :parser => %Membrane.H264.FFmpeg.Parser{
+              framerate: {60, 1}
+            },
+            {:sink, track_id} => %Membrane.File.Sink{
+              location: Path.join(state.output_dir, output_file)
+            }
+          }
+
+          links = [
+            link(:demuxer)
+            |> via_out(Pad.ref(:output, track_id))
+            |> to(:parser)
+            |> to({:sink, track_id})
+          ]
+
+          {{:ok, spec: %ParentSpec{children: children, links: links}}, state}
+
+        true ->
+          raise "Unsupported codec #{track_info.codec}"
       end
     end
   end
 
-  defp test_stream(input_file, references, tmp_dir) do
+  defp test_stream(input_file, track_id_to_reference, tmp_dir) do
     {:ok, pipeline} =
       %Testing.Pipeline.Options{
         module: TestPipeline,
         custom_args: %{
           input_file: Path.join(@fixtures_dir, input_file),
-          output_dir: tmp_dir
+          output_dir: tmp_dir,
+          track_id_to_output_file: track_id_to_reference
         }
       }
       |> Testing.Pipeline.start_link()
 
     Testing.Pipeline.play(pipeline)
     assert_pipeline_playback_changed(pipeline, _, :playing)
+
+    references = Map.values(track_id_to_reference)
 
     if Enum.count(references) == 1 do
       assert_end_of_stream(pipeline, {:sink, 1})
@@ -98,9 +127,24 @@ defmodule Membrane.WebM.DemuxerTest do
     Testing.Pipeline.stop_and_terminate(pipeline, blocking?: true)
     assert_pipeline_playback_changed(pipeline, _, :stopped)
 
-    for reference <- references do
-      assert File.read!(Path.join(@fixtures_dir, reference)) ==
-               File.read!(Path.join(tmp_dir, reference))
+    for reference <- Map.values(track_id_to_reference) do
+      reference_file = File.read!(Path.join(@fixtures_dir, reference))
+      result_file = File.read!(Path.join(tmp_dir, reference))
+
+      assert byte_size(reference_file) == byte_size(result_file), "Different size: #{reference}"
+
+      fixtures_list = :binary.bin_to_list(reference_file)
+      result_list = :binary.bin_to_list(result_file)
+
+      zipped_with_indexes = fixtures_list |> Enum.zip(result_list) |> Enum.with_index()
+
+      for {{elem1, elem2}, idx} = _elem <- zipped_with_indexes do
+        if elem1 != elem2 do
+          raise "#{elem1} is not equal #{elem2} on index #{idx}"
+        end
+      end
+
+      assert reference_file == result_file, "Different file: #{reference} "
     end
   end
 
@@ -113,17 +157,27 @@ defmodule Membrane.WebM.DemuxerTest do
   end
 
   @tag :tmp_dir
-  test "demuxing webm containing opus", %{tmp_dir: tmp_dir} do
-    test_stream("muxed_opus.webm", ["1.ogg"], tmp_dir)
+  test "demuxing mkv containing opus", %{tmp_dir: tmp_dir} do
+    test_stream("muxed_opus.mkv", %{1 => "1.ogg"}, tmp_dir)
   end
 
   @tag :tmp_dir
-  test "demuxing webm containing vp8 + opus", %{tmp_dir: tmp_dir} do
-    test_stream("vp8_opus_video.webm", ["1_vp8.ivf", "2.ogg"], tmp_dir)
+  test "demuxing mkv file (vp8,opus)", %{tmp_dir: tmp_dir} do
+    test_stream("vp8_opus_video.mkv", %{1 => "1_vp8_mkv.ivf", 2 => "2.ogg"}, tmp_dir)
   end
 
   @tag :tmp_dir
-  test "demuxing webm containing vp9 + opus", %{tmp_dir: tmp_dir} do
-    test_stream("vp9_opus_video.webm", ["1_vp9.ivf", "2.ogg"], tmp_dir)
+  test "demuxing muxed file", %{tmp_dir: tmp_dir} do
+    test_stream("combined_vp8.mkv", %{2 => "1_vp8.ivf", 1 => "1.ogg"}, tmp_dir)
+  end
+
+  @tag :tmp_dir
+  test "demuxing mkv file (vp9,opus)", %{tmp_dir: tmp_dir} do
+    test_stream("vp9_opus_video.mkv", %{1 => "1_vp9_mkv.ivf", 2 => "2.ogg"}, tmp_dir)
+  end
+
+  @tag :tmp_dir
+  test "demuxing mkv file (h264,opus)", %{tmp_dir: tmp_dir} do
+    test_stream("combined_h264.mkv", %{2 => "video.h264", 1 => "1.ogg"}, tmp_dir)
   end
 end
